pyproject.toml
[tool.poetry]
name = "hack-museum-service"
version = "0.1.0"
description = ""
authors = ["JKearnsl <boyarshinovn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
SQLAlchemy = "^2.0.19"
asyncpg = "^0.29.0"

aiohttp = "^3.9.4"
aiobotocore = "^2.7.0"

python-consul = "^1.1.0"
python-jose = { version = "==3.3.*", extras = ["cryptography"] }
grpcio = "^1.59.0"
APScheduler = "^3.10.4"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"
grpcio-tools = "^1.57.0"
types-aiobotocore = "^2.6.0"

PyYAML = "^6.0.1"
python-dotenv= "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
